[Refactor] Extract setting dialog logic into composables
